2021/2/28 原料名、会社名登録。原料在庫計算機能追加。

+++ database/migrations/2021_02_27_044904_create_materials_table.php
@@ -16,7 +16,7 @@ class CreateMaterialsTable extends Migration
         Schema::create('materials', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->char('name', 100);
-            $table->tinyInteger('company_id');
+            $table->integer('company_id');
             $table->timestamps();
         });
     }
